+++ lib/src/utils/logger.dart
@@ -1,0 +1,6 @@
+import 'package:logger/logger.dart';
+
+var logger = Logger(
+  printer: PrettyPrinter(methodCount: 2),
+  // printer: SimplePrinter(),
+);